% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_to_mm.R
\name{af_to_mm}
\alias{af_to_mm}
\title{acre-feet to mm streamflow conversion}
\usage{
af_to_mm(af, km2)
}
\arguments{
\item{af}{Acre-feet of streamflow}

\item{km2}{Watershed area in km2}
}
\description{
Converts streamflow in acre-feet to mm
}
\examples{
# Compute 12 inches (304.8 mm) of water (arguments are 1 acre-foot and 1 acre, respectively)
af_to_mm(1, 0.00404686)

}
\author{
Ryan Bart
}
