% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_patches_in_watershed.R
\name{calculate_patches_in_watershed}
\alias{calculate_patches_in_watershed}
\title{Calculate how many patches in a watershed}
\usage{
calculate_patches_in_watershed(
  watershed_area_km2 = NULL,
  watershed_area_mi2 = NULL,
  patch_side_m
)
}
\arguments{
\item{watershed_area_km2}{Watershed area in square kilometers. Dafaults to
this value if both watershed_area_mi2 and watershed_area_mi2 are included
as arguments.}

\item{watershed_area_mi2}{Watershed area in square miles.}

\item{patch_side_m}{Length of a patch side in meters}
}
\description{
Converts watershed size and patch size to number of patches
}
\examples{
# Compute number of 100m patches in a square km and square mile
calculate_patches_in_watershed(watershed_area_mi2 = 1, patch_side_m = 100)
calculate_patches_in_watershed(watershed_area_km2 = 1, patch_side_m = 100)

}
\author{
Ryan Bart
}
