% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfs_to_mm.R
\name{cfs_to_mm}
\alias{cfs_to_mm}
\title{cfs to mm/day streamflow conversion}
\usage{
cfs_to_mm(cfs, km2)
}
\arguments{
\item{cfs}{Flow in cubic feet per second}

\item{km2}{Watershed area in km2}
}
\description{
Converts streamflow in cubic feet per second to mm/day
}
\examples{
# Change streamflow at 1 cfs to mm/day for a 1 acre watershed
cfs_to_mm(1, 0.00404686)

}
\author{
Ryan Bart
}
