% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cy_to_wy.R
\name{cy_to_wy}
\alias{cy_to_wy}
\title{Calendar year to water year conversion}
\usage{
cy_to_wy(cy, m, start_month = 10)
}
\arguments{
\item{cy}{Calendar year}

\item{m}{Month as a integer (e.g. October = 10)}

\item{start_month}{Starting month as a integer (e.g. October = 10)}
}
\description{
Converts calendar year to water year based on month
}
\author{
Ryan Bart
}
