% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyd_to_m.R
\name{cyd_to_m}
\alias{cyd_to_m}
\title{Calendar-year day to month conversion}
\usage{
cyd_to_m(cyd, cy)
}
\arguments{
\item{cyd}{Calendar-year day}

\item{cy}{Calendar year}
}
\description{
Converts calendar-year day to month
}
\author{
Ryan Bart
}
