% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyd_to_wyd.R
\name{cyd_to_wyd}
\alias{cyd_to_wyd}
\title{calendar-year day to water-year day conversion}
\usage{
cyd_to_wyd(cyd, cy)
}
\arguments{
\item{cyd}{Calendar-year day}

\item{cy}{Calendar year}
}
\description{
Converts calendar-year day to water-year day based on a Oct-Sept water year.
}
\section{Note}{
 This function should be generalized for use with any start_day
  for water year
}

\author{
Ryan Bart
}
