% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_cyd.R
\name{date_to_cyd}
\alias{date_to_cyd}
\title{date to calendar-year day conversion}
\usage{
date_to_cyd(cy, m, d)
}
\arguments{
\item{cy}{Calendar year}

\item{m}{Month}

\item{d}{Day of month}
}
\description{
Converts a date (inputs of year, month, day) to calendar-year day
}
\author{
Ryan Bart
}
