% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_wyd.R
\name{date_to_wyd}
\alias{date_to_wyd}
\title{date to water-year day conversion}
\usage{
date_to_wyd(cy, cyd, wyd_start = 274)
}
\arguments{
\item{cy}{Calendar year}

\item{cyd}{Calendar-year day. January 1 is equal to 1. Can be obtained from
date_to_cyd function.}

\item{wyd_start}{Calendar-year day to start water-year day (Default 274 is
October 1)}
}
\description{
Converts a date (inputs of year, calendar-year day) to water-year day
}
\section{Note}{
 Accounts for leap years (but not the skipping of leap years at
  every century except those divisible by 400 (e.g. year 2000))

  Tip: Function requires purrr:pmap_dbl or mapply

  purrr::pmap_dbl(list(cy, cyd, wyd_start=274), date_to_wyd)

  mapply(date_to_wyd, cy, cyd, wyd_start)
}

\author{
Ryan Bart
}
