% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wy_to_cy.R
\name{wy_to_cy}
\alias{wy_to_cy}
\title{Water year to calendar year conversion}
\usage{
wy_to_cy(wy, m, start_month = 10)
}
\arguments{
\item{wy}{Water year}

\item{m}{Month as a integer (e.g. October = 10)}

\item{start_month}{Starting month as a integer (e.g. October = 10)}
}
\description{
Converts calendar year to water year based on month
}
\author{
Ryan Bart
}
