% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyd_to_cyd.R
\name{wyd_to_cyd}
\alias{wyd_to_cyd}
\title{water-year day to calendar-year day conversion}
\usage{
wyd_to_cyd(wyd, cy)
}
\arguments{
\item{wyd}{Water-year day}

\item{cy}{Calendar year}
}
\description{
Converts water-year day to calendar-year day based on a Oct-Sept water year.
}
\section{Note}{
 This function should be generalized for use with any start_day
  for water year
}

\author{
Ryan Bart
}
