% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_to_wy.R
\name{y_to_wy}
\alias{y_to_wy}
\title{Calendar year to water year conversion}
\usage{
y_to_wy(year, month, start_month = 10)
}
\arguments{
\item{year}{Calendar year}

\item{month}{Month Month as a integer (e.g. October = 10)}

\item{start_month}{Starting month as a integer (e.g. October = 10)}
}
\description{
Converts calendar year to water year based on month
}
\author{
Ryan Bart
}
