% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_to_af.R
\name{mm_to_af}
\alias{mm_to_af}
\title{mm/day to acre-feet streamflow conversion}
\usage{
mm_to_af(mm, km2)
}
\arguments{
\item{mm}{Millimeters of streamflow per day}

\item{km2}{Watershed area in km2}
}
\description{
Converts streamflow at mm/day to acre-feet
}
\examples{
# Compute 1 acre-foot of water (arguments are 1 foot and 1 acre, respectively)
mm_to_af(25.4*12, 0.00404686)

}
\author{
Ryan Bart
}
